Refactor: 추가 컴포넌트화
