+++ src/basic/components/Cart.tsx
@@ -1,7 +1,7 @@
 // src/basic/components/Cart.tsx
 import { CartItem, Coupon } from '../types';
-import { formatCurrency } from '../utils/formatters';
-import { getMaxApplicableDiscount } from '../models/cart';
+import { CartItemCard } from './CartItemCard';
+import { CartSummary } from './CartSummary';
 
 interface Props {
   cart: CartItem[];
@@ -47,50 +47,14 @@ export const Cart = ({
           </div>
         ) : (
           <div className="space-y-3">
-            {cart.map(item => {
-              const discount = getMaxApplicableDiscount(item);
-              const discountRateText = `-${Math.round(discount.rate * 100)}%`;
-              return (
-                <div key={item.id} className="border-b pb-3 last:border-b-0">
-                  <div className="flex justify-between items-start mb-2">
-                    <h4 className="text-sm font-medium text-gray-900 flex-1">{item.name}</h4>
-                    <button
-                      onClick={() => onRemoveFromCart(item.id)}
-                      className="text-gray-400 hover:text-red-500 ml-2"
-                    >
-                      <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-                      </svg>
-                    </button>
-                  </div>
-                  <div className="flex items-center justify-between">
-                    <div className="flex items-center">
-                      <button
-                        onClick={() => onUpdateQuantity(item.id, item.quantity - 1)}
-                        className="w-6 h-6 rounded border border-gray-300 flex items-center justify-center hover:bg-gray-100"
-                      >
-                        <span className="text-xs">−</span>
-                      </button>
-                      <span className="mx-3 text-sm font-medium w-8 text-center">{item.quantity}</span>
-                      <button
-                        onClick={() => onUpdateQuantity(item.id, item.quantity + 1)}
-                        className="w-6 h-6 rounded border border-gray-300 flex items-center justify-center hover:bg-gray-100"
-                      >
-                        <span className="text-xs">+</span>
-                      </button>
-                    </div>
-                    <div className="text-right">
-                      {discount.rate > 0 && (
-                        <span className="text-xs text-red-500 font-medium block">{discountRateText}</span>
-                      )}
-                      <p className="text-sm font-medium text-gray-900">
-                        {formatCurrency(item.price * item.quantity)}
-                      </p>
-                    </div>
-                  </div>
-                </div>
-              )
-            })}
+            {cart.map(item => (
+              <CartItemCard 
+                key={item.id}
+                item={item}
+                onUpdateQuantity={onUpdateQuantity}
+                onRemoveFromCart={onRemoveFromCart}
+              />
+            ))}
           </div>
         )}
       </section>
@@ -123,36 +87,7 @@ export const Cart = ({
             )}
           </section>
 
-          <section className="bg-white rounded-lg border border-gray-200 p-4">
-            <h3 className="text-lg font-semibold mb-4">결제 정보</h3>
-            <div className="space-y-2 text-sm">
-              <div className="flex justify-between">
-                <span className="text-gray-600">상품 금액</span>
-                <span className="font-medium">{formatCurrency(cartTotal.totalBeforeDiscount)}</span>
-              </div>
-              {(cartTotal.totalDiscount + cartTotal.couponDiscount) > 0 && (
-                <div className="flex justify-between text-red-500">
-                  <span>할인 금액</span>
-                  <span>-{formatCurrency(cartTotal.totalDiscount + cartTotal.couponDiscount)}</span>
-                </div>
-              )}
-              <div className="flex justify-between py-2 border-t border-gray-200">
-                <span className="font-semibold">결제 예정 금액</span>
-                <span className="font-bold text-lg text-gray-900">{formatCurrency(cartTotal.finalTotal)}</span>
-              </div>
-            </div>
-
-            <button
-              onClick={onCheckout}
-              className="w-full mt-4 py-3 bg-yellow-400 text-gray-900 rounded-md font-medium hover:bg-yellow-500 transition-colors"
-            >
-              {formatCurrency(cartTotal.finalTotal)} 결제하기
-            </button>
-
-            <div className="mt-3 text-xs text-gray-500 text-center">
-              <p>* 실제 결제는 이루어지지 않습니다</p>
-            </div>
-          </section>
+          <CartSummary cartTotal={cartTotal} onCheckout={onCheckout} />
         </>
       )}
     </div>
